Introduce system dir, non entity call will move here, detached Versions calls to VersionsMixin
