+++ gvm/protocols/gmpv208/__init__.py
@@ -139,9 +139,10 @@ from gvm.protocols.gmpv208.entities.users import (
     get_user_auth_type_from_string,
 )
 from gvm.protocols.gmpv208.entities.vulnerabilities import VulnerabilitiesMixin
-from gvm.connections import GvmConnection
 
-PROTOCOL_VERSION = (20, 8)
+from gvm.protocols.gmpv208.system.version import VersionMixin
+
+from gvm.connections import GvmConnection
 
 
 class Gmp(
@@ -178,6 +179,7 @@ class Gmp(
     SecInfoMixin,
     UserSettingsMixin,
     UsersMixin,
+    VersionMixin,
     VulnerabilitiesMixin,
 ):
     """Python interface for Greenbone Management Protocol
@@ -210,12 +212,3 @@ class Gmp(
 
         # Is authenticated on gvmd
         self._authenticated = False
-
-    @staticmethod
-    def get_protocol_version() -> tuple:
-        """Determine the Greenbone Management Protocol version.
-
-        Returns:
-            tuple: Implemented version of the Greenbone Management Protocol
-        """
-        return PROTOCOL_VERSION
