+++ gvm/protocols/gmpv214/system/version.py
@@ -1,0 +1,36 @@
+# -*- coding: utf-8 -*-
+# Copyright (C) 2021 Greenbone Networks GmbH
+#
+# SPDX-License-Identifier: GPL-3.0-or-later
+#
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with this program.  If not, see <http://www.gnu.org/licenses/>.
+
+
+from gvm.protocols.gmpv208.system.version import (
+    VersionMixin as Gmp208VersionMixin,
+)
+
+PROTOCOL_VERSION = (21, 4)
+
+
+class VersionMixin(Gmp208VersionMixin):
+    @staticmethod
+    def get_protocol_version() -> tuple:
+        """Determine the Greenbone Management Protocol (gmp) version used
+        by python-gvm version.
+
+        Returns:
+            tuple: Implemented version of the Greenbone Management Protocol
+        """
+        return PROTOCOL_VERSION
